Built the base structure for the media-downloader (#1)
Base structure allows the user to input a file location; '~/' gets
interpreted as '/home/<OsUsername>/'. They're allowed to exit the
program during the prompting for a url, they're also allowed to input
the format of the media they want. Currently I want to solve the error
of a media name being too long that program aborts it and prompts for a
url once more.